# VM migration: Reduce downtime using LSP IFSTATE