#### What this PR does and why is it needed
Ccurrently all pods have the LSP.Enabled option set to true by default.

During migration, it makes sense to handoff the network traffic from the
source to target pod as soon as the target pod is up and ready for
traffic.
This PR is performing this handoff by closing the LSP of the source port.
Doing this significantly reduces the migration downtime.

This traffic handoff optimization is currently restricted to VM with
only secondary interfaces, and layer2 topology.

#### Which issue(s) this PR fixes

Fixes #

#### Special notes for reviewers

#### How to verify it
Run migration tests.

#### Details to documentation updates
Migration downtime results:
pre-copy:
  before change: 13s
  After change: 3s
post-copy:
  before change: 26s
  After change: 13s

#### Description for the changelog

#### Does this PR introduce a user-facing change?
```release-note
NONE
```