@qinqon 2024-10-15
> /cc @qinqon
> 
> Could you take this for a drive, and check if this matches your expectations ?

sure.