@RamLavi 2024-10-30
[Change](https://github.com/ovn-org/ovn-kubernetes/compare/f0945f0b86d2736bceece3bf4a907f6c01d662c2..b08fe373ca24a6974e4bb7ee4614da3b3320757f): Small fix to lsp logic, making the Failed migration happen fast