@RamLavi 2024-11-11
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/55d871388d10e72592236d8d6d82733430d0c3db..a76bdb55f5758f132deb5f089450cdeae0ad93e3): Address @qinqon 's reviews, and also add localnet scenarios in the appropriate commit