@RamLavi 2024-11-12
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/a60d1e1f533349e12d4ab31e0da5d3536fecfdde..209e03b909122f501463ac3918a2c2af81c32beb): Removing unit tests failMigration on nonIC cases (ovn annotations are not supported on test infra, causing the test to not pass)