@RamLavi 2024-11-12
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/81d584b6a4ac35e98c27e16dff686fd86ae4c311..7603811fbc6b2bb2d2fbb0e95e03f4f08789cf6f): Moved some test code to helper to make test more readable.