@RamLavi 2024-11-12
@maiqueb I tried to refactor the `expectedLogicalSwitchesAndPorts()` function some more so that I wouldn't need the second map `alreadyAddedManagementElements`... the params are pretty tangled up so it's not trivial to get something nice and simple out of it. Tell me if this still bugs you and I'll give it another go.