@RamLavi 2024-11-18
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/cbb2fa054463468d3555eeaff01e01212edcaac8..4d37a887f537309634ab2c57c0882d5d4fba77c1): fixed of recent reviews from @qinqon and @maiqueb