@RamLavi 2024-11-18
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/c952d070ddd78fd2a2359f9136be8ec8e757021e..6897fc780f0b4d635f0916df8049d081b4a98c6d): Refactor code outside of BaseNetworkController.