@qinqon 2024-11-19
@RamLavi can you add the following link to PR's description to show where the e2e tests do live.
- https://github.com/ovn-kubernetes/ovn-kubernetes/pull/4834