@RamLavi 2024-11-19
[Changes](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/d3a127b3b65aebbd64ef234b8b2c9ff04af5925a..d370577a838e28be71902c5c337e039aae3e56e6): address @qinqon 's review