@RamLavi 2024-11-19
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/d370577a838e28be71902c5c337e039aae3e56e6..2391c94d3b0e483d6b3f0da14096502714602924): address @maiqueb review