@RamLavi 2024-11-20
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/2391c94d3b0e483d6b3f0da14096502714602924..100e410efe9b8ff6be50f003949c4cfccf57026a): address @qinqon's review