@RamLavi 2024-11-21
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/100e410efe9b8ff6be50f003949c4cfccf57026a..61fe1403bd8c1f0d0f36b8fb56607292f9e87dc7): fix @maiqueb 's review