@RamLavi 2024-11-26
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/1d6fc2993f9b51d5628efcc0e7ccbbdf31585047..48cdf27497ec3f56e69b5633fe66096d7b701995): Addressed @kyrtapz 's review