@RamLavi 2024-11-27
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/dc73fb073be905a0fc814e71479956628bdac44a..257fe1c7f7f4a1414b12e670ea54b23a1d582fc9): Implemented `getNonZeroFields` function in order to not set the Enabled field every time when not needed.
@kyrtapz PTAL