@RamLavi 2024-12-03
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/f7252a765a64f808c168fbd60eb2f10d073b85ae..8703ce0379546db79fdcd8d42064e3207b79d5ce): Fix @jcaamano's review