@RamLavi 2024-12-03
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/8703ce0379546db79fdcd8d42064e3207b79d5ce..1ef70b6dbe945653ac30d7aee15ff6613655267f): small fix accounting also Option changes made by `AddTransitPortConfig`