@RamLavi 2024-12-03
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/1ef70b6dbe945653ac30d7aee15ff6613655267f..414d3e5454164ced40cde6f78213c091706ebe73): fix @jcaamano 's review