@RamLavi 2024-12-04
[Change](https://github.com/ovn-kubernetes/ovn-kubernetes/compare/c790b3614065d288f4bd0c4f9043af133ed6c122..479446d0c3777bd501693cd6cc4264080f2f73b1): Address @jcaamano 's remove