@RamLavi 2024-12-05
/hold 
kubevirt persistent-ip tests failed: checking this locally now