@RamLavi 2024-12-05
> /hold kubevirt persistent-ip tests failed: checking this locally now

The current solution was broken after https://github.com/ovn-kubernetes/ovn-kubernetes/pull/4848 was merged, as the network controller now ignores cluster-networked pods that do not want a port configured(addPort==false), so when target pod becomes ready the LSP.Enabled won't change...

It seems like this PR introduces the first scenario where the port needs to be updated even though  it was already configured (addPort==false)
@kyrtapz do you have a suggestion how to handle this?